Rename apache files.
This project is intended to run in a vm, and
several vm's can be run on a single web server,
so there should never be any need to run more
than one site in this vm and thus no need to
have specific names for apache site files.

If there's ever a need for that, we can rename
these files then.